リファクタリング 不要な関数を削除 関数のservicesへの切り出し
