+++ src/components/Search.tsx
@@ -1,4 +1,4 @@
-import React, { useContext, useState } from "react";
+import React, { useContext } from "react";
 import { changeSearchValue } from "../actions/ActionCreator";
 import { Context } from "../contexts/Context";
 import { omdbFetch } from "../services/omdb";
