+++ src/services/isBottom.ts
@@ -1,0 +1,3 @@
+export const isBottom = () => {
+  return window.scrollY + window.innerHeight + 1 >= document.body.scrollHeight;
+};
