+++ src/services/isMovieObject.ts
@@ -1,0 +1,11 @@
+import { MovieObject } from "../components/App";
+
+export const isMovieObject = (obj: any): obj is MovieObject => {
+  return (
+    typeof obj.Title === "string" &&
+    typeof obj.Type === "string" &&
+    typeof obj.Poster === "string" &&
+    typeof obj.Year === "string" &&
+    typeof obj.imdbID === "string"
+  );
+};
